gh-71966: Move the module docstring from _pydecimal to decimal
Now it is set even if the C implementation is used.

Also add a one-line synopsis.
